refactor(utils): replace copy_rtl_files with generate_sources_list
- Remove copy_rtl_files.cmake and copy_rtl_files.py utilities
- Add generate_sources_list.cmake and generate_sources_list.py using slang
- Move and update read_rtl_sources.cmake for new source list format
- Update SoCMakeConfig.cmake and tests to use new utilities
