+++ cmake/utils/generate_sources_list/read_rtl_sources.cmake
